+++ utils/utils.go
@@ -20,6 +20,8 @@ const (
 	ErrBadAbsLocationPath = "absolute location path is invalid - must include leading and trailing slashes"
 	// ErrBadRelLocationPath constant is returned when a file path is not relative
 	ErrBadRelLocationPath = "relative location path is invalid - may not include leading slash but must include trailing slash"
+	//TouchCopyBuffered min buffer size in bytes
+	TouchCopyMinBufferSize = 262144
 )
 
 // regex to test whether the last character is a '/'
@@ -96,10 +98,41 @@ func EnsureLeadingSlash(dir string) string {
 	return "/" + dir
 }
 
+// Deprecated - Use TouchCopyBuffer Instead
 // TouchCopy is a wrapper around io.Copy which ensures that even empty source files (reader) will get written as an
 // empty file. It guarantees a Write() call on the target file.
 func TouchCopy(writer io.Writer, reader io.Reader) error {
 	size, err := io.Copy(writer, reader)
+	if err != nil {
+		return err
+	}
+	if size == 0 {
+		_, err = writer.Write([]byte{})
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// TouchCopyBuffer is a wrapper around io.CopyBuffer which ensures that even empty source files (reader) will get written as an
+// empty file. It guarantees a Write() call on the target file.
+// bufferSize is in bytes and if is less than TouchCopyMinBufferSize will result in a buffer of size TouchCopyMinBufferSize
+// bytes. If bufferSize is > TouchCopyMinBufferSize it will result in a buffer of size bufferSize bytes
+func TouchCopyBuffered(writer io.Writer, reader io.Reader, bufferSize int) error {
+	var buffer []byte
+	var size int64
+	var err error
+
+	if bufferSize > TouchCopyMinBufferSize {
+		buffer = make([]byte, bufferSize)
+		size, err = io.CopyBuffer(writer, reader, buffer)
+	} else {
+		//Use Default Buffer of 256KB (256 * 1024 Bytes)
+		buffer = make([]byte, TouchCopyMinBufferSize)
+		size, err = io.CopyBuffer(writer, reader, buffer)
+	}
+
 	if err != nil {
 		return err
 	}